requirements.txt
opencv-python-headless
picamera
RPi.GPIO
adafruit-circuitpython-servokit
imutils
Flask
